requirements.txt
annotated-types==0.7.0
anyio==4.11.0
bcrypt==5.0.0
click==8.3.0
colorama==0.4.6
ecdsa==0.19.1
fastapi==0.118.0
greenlet==3.2.4
h11==0.16.0
httptools==0.6.4
idna==3.10
passlib==1.7.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pydantic==2.11.9
pydantic_core==2.33.2
python-dotenv==1.1.1
python-jose==3.5.0
PyYAML==6.0.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
unicorn==2.1.4
uvicorn==0.37.0
watchfiles==1.1.0
websockets==15.0.1
